Rename Onslaught → Empire and tune for full-pool rankings
Three design refinements over the v1 prototype, each suggested by play
feedback:

  - Drop the suicide-strength gate. The 1.4x attacker bonus makes the
    trade favorable at any strength: an attack that loses still deals
    more damage than the attacker gives up, and the soft survivor falls
    to the next attacker on the same border. Sitting still on a
    saturated border is worse than any 1.4x trade.

  - Focus enemy. When several pacifists surround us symmetrically, our
    armies attacking three different turtles convert to zero kills.
    Empire picks the living non-self player with the least territory
    each tick and gives any tile they own a +600 score bonus, so all
    armies pile pressure on one player at a time. Eliminations
    compound — fewer rivals leaves more uncontested space for
    everyone, including us.

  - Cap discipline. A march through a friendly tile only fires if the
    friendly has headroom, and the commitment is capped at that
    headroom. Pumping attackPower into a maxed-out neighbor clamps to
    zero growth; sitting at cap costs ~0 while a wasted transfer
    bleeds the source.

Officially seated on the board: 5 seasons run on the 51-bot full pool
(rankings.json refreshed). Empire ranks 64/64 in the full pool because
the survivor-friendly Pacifist/Cluster generation overweights
average-finish; in 14-bot core-bot subsets Empire is mid-pack to
top-tier (51% PL, often the rating leader). Its niche is
stalemate-breaking against turtles, not dominant winning.
