+++ src/strategies/index.js
@@ -37,7 +37,7 @@ import Pinwheel from "./Pinwheel.js";
 import Anvil from "./Anvil.js";
 import Skirmisher from "./Skirmisher.js";
 import Tempo from "./Tempo.js";
-import Onslaught from "./Onslaught.js";
+import Empire from "./Empire.js";
 import { GENERATED } from "./generated.js";
 import { DESCENDANTS } from "./descendants.js";
 import { ARCHIVED } from "./archive.js";
@@ -85,7 +85,7 @@ export const ALL_STRATEGY_LIST = [
   Anvil,
   Skirmisher,
   Tempo,
-  Onslaught,
+  Empire,
   ...GENERATED,
   ...DESCENDANTS,
 ];
